Split PhotoDetailHeader into two components
