+++ components/General/PhotoDetailHeader.js
@@ -1,23 +1,14 @@
 import { Query } from '@apollo/react-components';
-import Button from '@material-ui/core/Button';
-import Tooltip from '@material-ui/core/Tooltip';
-import Typography from '@material-ui/core/Typography';
-import MapIcon from '@material-ui/icons/Map';
 import React, { Fragment, useEffect, useState } from 'react';
 import { isWebpSupported } from 'react-image-webp/dist/utils';
-import ProgressiveImage from 'react-progressive-image';
-import { getBudgetScore } from '../../helpers/budgetScore';
 import { nameFromSlug } from '../../helpers/countryCodes';
 import { imageProxy } from '../../helpers/getImage';
 import { GET_LOCATION_DETAILS } from '../../helpers/graphql/locations';
-import Link from '../../lib/Link';
-import PopularDestinationsPopup from '../Destinations/PopularDestinationsPopup';
-import Head from '../Header/Head';
-import EditLocationDetails from './EditLocationDetails';
-import TopicSelector from './TopicSelector';
+import PhotoHeader from './PhotoHeader';
 
 const PhotoDetailHeader = props => {
   const { query, countrySlug, title, tag } = props;
+
   const [screenWidth, setScreenWidth] = useState(1920);
   const [webpSupport, setWebpSupport] = useState(undefined);
 
@@ -32,32 +23,33 @@ const PhotoDetailHeader = props => {
     <Fragment>
       <Query query={GET_LOCATION_DETAILS} variables={query}>
         {({ data, loading }) => {
-          const detailTitle =
-            (data && data.locationDetails && data.locationDetails.title) ||
-            title;
-          let description = `${
-            tag ? `Discover ${detailTitle}` : `Visit ${detailTitle}`
-          }: ${
-            data &&
-            data.locationDetails.budget_score &&
-            data.locationDetails.budget_score < 3
-              ? `Traveling ${detailTitle} is affordable and exciting! `
-              : ''
-          }`;
-          if (
-            data &&
-            !data.locationDetails.url &&
-            data.locationDetails.description
-          )
-            description += `${data.locationDetails.description} `;
-          description += `Explore the best places to visit, discover insider's tips, find out what to do${
-            data &&
-            data.locationDetails.budget_score &&
-            data.locationDetails.budget_score > 4
-              ? ` in ${detailTitle} on a budget`
-              : ''
-          } and read the top travel blogs about ${detailTitle} on TravelFeed!`;
           if (loading || (data && data.locationDetails)) {
+            const detailTitle =
+              (data && data.locationDetails && data.locationDetails.title) ||
+              title;
+            let description = `${
+              tag ? `Discover ${detailTitle}` : `Visit ${detailTitle}`
+            }: ${
+              data &&
+              data.locationDetails.budget_score &&
+              data.locationDetails.budget_score < 3
+                ? `Traveling ${detailTitle} is affordable and exciting! `
+                : ''
+            }`;
+            if (
+              data &&
+              !data.locationDetails.url &&
+              data.locationDetails.description
+            )
+              description += `${data.locationDetails.description} `;
+            description += `Explore the best places to visit, discover insider's tips, find out what to do${
+              data &&
+              data.locationDetails.budget_score &&
+              data.locationDetails.budget_score > 4
+                ? ` in ${detailTitle} on a budget`
+                : ''
+            } and read the top travel blogs about ${detailTitle} on TravelFeed!`;
+
             let image = '';
             if (data && data.locationDetails.unsplashUser) {
               image = webpSupport
@@ -73,301 +65,22 @@ const PhotoDetailHeader = props => {
               );
             }
 
-            const innerCard = (
-              <>
-                <div className="container h-100">
-                  <div
-                    className="row h-100 justify-content-center"
-                    style={{ minHeight: '400px' }}
-                  >
-                    <div className="col-xl-7 col-lg-8 col-md-9 col-sm-10 col-12 my-auto">
-                      <Typography
-                        variant="h6"
-                        align="center"
-                        className="text-light font-weight-bold"
-                        style={{
-                          textShadow: '1px 1px 20px #343A40',
-                        }}
-                        component="h5"
-                      >
-                        {// For cities, display breadcrumbs to Country and Subdivison. Else, display Knowledge Graph subtitle, if unavailable country name
-                        (tag && `#${tag}`) ||
-                          (query.city && (
-                            <span>
-                              <Link
-                                color="textPrimary"
-                                as={`/destinations/${countrySlug}/`}
-                                href="/destinations/[...destination]"
-                              >
-                                <span className="text-light font-weight-bold">
-                                  {countryName}
-                                </span>
-                              </Link>
-                              <span className="text-light">
-                                {' '}
-                                &raquo;{' '}
-                                <Link
-                                  color="textPrimary"
-                                  as={`/destinations/${countrySlug}/${query.subdivision}`}
-                                  href="/destinations/[...destination]"
-                                >
-                                  <span className="text-light font-weight-bold">
-                                    {query.subdivision}
-                                  </span>
-                                </Link>
-                              </span>{' '}
-                            </span>
-                          )) ||
-                          (((query.search && !query.country_code) ||
-                            query.subdivision ||
-                            query.city) && (
-                            <Link
-                              color="textPrimary"
-                              as={`/destinations/${countrySlug}/`}
-                              href="/destinations/[...destination]"
-                            >
-                              <span className="text-light font-weight-bold">
-                                {(data && data.locationDetails.subtitle) ||
-                                  countryName}
-                              </span>
-                            </Link>
-                          ))}
-                        {!props.noEdit && (
-                          <EditLocationDetails
-                            country_code={props.query.country_code}
-                            subdivision={props.query.subdivision}
-                            city={props.query.city}
-                            tag={tag}
-                            data={
-                              data && data.locationDetails
-                                ? data.locationDetails
-                                : {}
-                            }
-                          />
-                        )}
-                      </Typography>
-                      <Typography
-                        gutterBottom
-                        variant="h2"
-                        component="h1"
-                        className="text-light font-weight-bold text-center"
-                        style={{
-                          textShadow: '1px 1px 10px #343A40',
-                        }}
-                      >
-                        {detailTitle}
-                      </Typography>
-                      <Typography
-                        gutterBottom
-                        className="lead text-light text-center"
-                        variant="h6"
-                        style={{
-                          textShadow: '1px 1px 10px black',
-                        }}
-                      >
-                        <em>
-                          {(data && data.locationDetails.description) || (
-                            <>
-                              <br />
-                              <br />
-                            </>
-                          )}
-                        </em>
-                      </Typography>
-                      <div className="text-center">
-                        {!tag && (
-                          <>
-                            <Link href={`/map?search=${title}`}>
-                              <Button
-                                className="m-2"
-                                variant="contained"
-                                color="primary"
-                              >
-                                <span className="pr-1">Explore the map</span>
-                                <MapIcon />
-                              </Button>
-                            </Link>
-                          </>
-                        )}
-                        {data &&
-                          data.locationDetails.sublocations &&
-                          data &&
-                          data.locationDetails.sublocations.length > 1 && (
-                            <PopularDestinationsPopup
-                              title={detailTitle}
-                              countrySlug={countrySlug}
-                              subdivision={query.subdivision}
-                              destinations={data.locationDetails.sublocations}
-                            />
-                          )}
-                        {data && data.locationDetails.budget_score && (
-                          <Tooltip
-                            title={getBudgetScore(
-                              data.locationDetails.budget_score,
-                              detailTitle,
-                            )}
-                          >
-                            <Typography
-                              component="p"
-                              variant="h4"
-                              className="pt-2 text-light font-weight-bold"
-                            >
-                              {'$'.repeat(data.locationDetails.budget_score)}
-                            </Typography>
-                          </Tooltip>
-                        )}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div
-                  className="text-mutedlight text-right pr-1"
-                  style={{ fontSize: '0.8rem' }}
-                >
-                  {data && data.locationDetails.url && (
-                    <span>
-                      Description by{' '}
-                      <a
-                        className="text-mutedlight text-decoration-underline"
-                        target="_blank"
-                        rel="nofollow noreferrer noopener"
-                        href={data && data.locationDetails.url}
-                      >
-                        Wikipedia
-                      </a>{' '}
-                      under{' '}
-                      <a
-                        className="text-mutedlight text-decoration-underline"
-                        target="_blank"
-                        rel="nofollow noreferrer noopener"
-                        href={data && data.locationDetails.license}
-                      >
-                        CC BY-SA 3.0
-                      </a>
-                      .{' '}
-                    </span>
-                  )}
-                  {data && data.locationDetails.attribution && (
-                    <span>
-                      Photo:{' '}
-                      {(data && data.locationDetails.unsplashUser && (
-                        <a
-                          className="text-mutedlight text-decoration-underline"
-                          target="_blank"
-                          rel="nofollow noreferrer noopener"
-                          href={`https://unsplash.com/@${data &&
-                            data.locationDetails
-                              .unsplashUser}?utm_source=TravelFeed&utm_medium=referral`}
-                        >
-                          {data && data.locationDetails.attribution}
-                        </a>
-                      )) || (
-                        <span className="text-mutedlight">
-                          {data && data.locationDetails.attribution}
-                        </span>
-                      )}{' '}
-                      {data && data.locationDetails.unsplashUser && (
-                        <span>
-                          /{' '}
-                          <a
-                            target="_blank"
-                            rel="nofollow noreferrer noopener"
-                            href="https://unsplash.com/?utm_source=TravelFeed&utm_medium=referral"
-                            className="text-mutedlight text-decoration-underline"
-                          >
-                            Unsplash
-                          </a>
-                        </span>
-                      )}
-                    </span>
-                  )}
-                </div>
-              </>
-            );
-            const loadedCard = (
-              <div
-                className="w-100"
-                style={{
-                  backgroundImage: `linear-gradient(rgba(0, 0, 0, 0.3),rgba(0, 0, 0,0.5)),
-      url("${image}")`,
-                  backgroundRepeat: 'no-repeat',
-                  backgroundPosition: 'center center',
-                  backgroundSize: 'cover',
-                }}
-              >
-                {innerCard}
-              </div>
-            );
-            const { country_code, subdivision, city } = props.query;
-            const breadcrumbs = [];
-            if (country_code) {
-              breadcrumbs.push(
-                {
-                  position: 1,
-                  name: 'Destinations',
-                  item: `https://travelfeed.io/destinations`,
-                },
-                {
-                  position: 2,
-                  name: countryName,
-                  item: `https://travelfeed.io/destinations/${props.countrySlug}`,
-                },
-              );
-            }
-            if (subdivision) {
-              breadcrumbs.push({
-                position: 3,
-                name: subdivision,
-                item: `https://travelfeed.io/destinations/${props.countrySlug}/${subdivision}`,
-              });
-            }
-            if (city) {
-              breadcrumbs.push({
-                position: 4,
-                name: city,
-                item: `https://travelfeed.io/destinations/${props.countrySlug}/${subdivision}/${city}`,
-              });
-            }
             return (
-              <Fragment>
-                <Head
-                  breadcrumbs={breadcrumbs}
-                  noIndex={props.noIndex}
-                  shorttitle={`${detailTitle}: Top Travel Blogs`}
-                  description={description}
-                />
-                {data && data.locationDetails && !data.locationDetails.image ? (
-                  loadedCard
-                ) : (
-                  <ProgressiveImage src={image} placeholder="">
-                    {(src, imgloading) => {
-                      if (imgloading) {
-                        return (
-                          <>
-                            <div
-                              className="MuiSkeleton-rect w-100 MuiSkeleton-animate"
-                              style={{
-                                minHeight: '400px',
-                                backgroundColor: 'rgba(0, 0, 0, 0.08)',
-                              }}
-                            >
-                              {innerCard}
-                            </div>
-                          </>
-                        );
-                      }
-                      return loadedCard;
-                    }}
-                  </ProgressiveImage>
-                )}
-                {data && data.locationDetails.topics && (
-                  <TopicSelector
-                    topics={data.locationDetails.topics}
-                    active={props.topic}
-                    onChange={props.setTopic}
-                  />
-                )}
-              </Fragment>
+              <PhotoHeader
+                title={detailTitle}
+                description={description}
+                image={image}
+                query={query}
+                countrySlug={countrySlug}
+                tag={tag}
+                subtitle={
+                  (data && data.locationDetails.subtitle) || countryName
+                }
+                countryName={countryName}
+                locationDetails={
+                  data && data.locationDetails ? data.locationDetails : {}
+                }
+              />
             );
           }
           return <Fragment />;
